requirements.txt
boto3==1.34.23
botocore==1.34.23
cryptography==41.0.3
folium==0.14.0
geopandas==0.14.0
httpx-oauth==0.13.2
pandas==1.5.3
Pillow==10.0.1
plotly==5.17.0
shapely==2.0.1
streamlit-vertical-slider==2.5.5
streamlit==1.34.0
streamlit_extras==0.4.0
streamlit_folium==0.15.0
geemap==0.32.0
ee==0.2
scipy==1.13.0
streamlit-extras==0.4.0